Extract tests module to file in ide_db crate
